Refactor: Movement (And Organize Files)
