+++ Assets/Scripts/Runtime/CH1/Pacmom/Dust.cs
@@ -10,7 +10,6 @@ namespace Runtime.CH1.Pacmom
 
         private void Start()
         {
-            movement.isRotationNeeded = false;
             ResetState();
         }
