+++ Assets/Scripts/Runtime/CH1/Pacmom/Movements.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40c49d28016f0144fbbe943b29418da5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
