+++ Assets/Scripts/Runtime/CH1/Pacmom/Movements/MovementAndEyes.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Runtime.CH1.Pacmom
+{
+    public class MovementAndEyes : Movement
+    {
+        private void Awake()
+        {
+            Set();
+        }
+    }
+}
