+++ Assets/Scripts/Runtime/CH1/Pacmom/Movements/MovementAndEyes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e52d0147f893494c8c45433a156e479
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
