+++ Assets/Scripts/Runtime/CH1/Pacmom/Movements/MovementAndRotation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bba9badf88412aa4fb2a67e5461e47a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
