+++ Assets/Scripts/Runtime/CH1/Pacmom/Movements/SpriteRotation.cs.meta
