+++ Assets/Scripts/Runtime/CH1/Pacmom/SpriteControls.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 676e01154dcc03b4c84461c25b19deb8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
